5_SQLQuery/src/requirements.txt
locust>=2.17.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0








